refactor: extract guardian text engine and legacy fallback strategies
